chnaged /history command
